refactor: rename node to element
