+++ packages/core/src/V2/nodes/link.ts
@@ -1,9 +1,9 @@
 import { OperatorElementConfig, PortProvider } from '../circuitElement';
 import { EMPTY, Observable } from 'rxjs';
 import { Operator } from './operator';
-import { CreateOutputPort } from './nodePorts';
+import { CreateOutputPort } from './elementPorts';
 
-export class LinkNodePorts implements PortProvider {
+export class LinkElementPorts implements PortProvider {
   constructor(private fromPort: Observable<unknown>, private toPortName: string) {
   }
 
@@ -19,7 +19,7 @@ export const Link: OperatorElementConfig = {
   boot: (param: unknown) => {
     const { from, to } = param as {from: string, to: string};
     const createLinkOutput: CreateOutputPort = (input) => {
-      return new LinkNodePorts(input.getPort(from), to);
+      return new LinkElementPorts(input.getPort(from), to);
     }
 
     return new Operator(createLinkOutput);
