+++ packages/core/src/V2/nodes/linkCount.ts
@@ -1,7 +1,7 @@
 // inputvalue === outputvalue
 // show the value pass numbe r of links
 
-import { CreateOutputPort, NodePorts } from './nodePorts';
+import { CreateOutputPort, ElementPorts } from './elementPorts';
 import { map, tap } from 'rxjs/operators';
 import { OperatorElementConfig } from '../circuitElement';
 import { Operator } from './operator';
@@ -10,7 +10,7 @@ export const LinkCount: OperatorElementConfig = {
   boot: (param: unknown) => {
     const linkCountParams = param as { getLinkCount: (count: number) => void };
     const createLinkCountOutput: CreateOutputPort = (input) => {
-      return new NodePorts(input.getPort('input').pipe(
+      return new ElementPorts(input.getPort('input').pipe(
         map((val, i) => {
           linkCountParams.getLinkCount(i + 1);
           return val;
