+++ packages/core/src/V2/nodes/sleep.test.ts
@@ -1,7 +1,7 @@
 import { of } from 'rxjs';
 import { TestScheduler } from 'rxjs/testing';
 import { Sleep } from './sleep';
-import { LinkNodePorts } from './link';
+import { LinkElementPorts } from './link';
 
 describe('sleep', () => {
   let testScheduler: TestScheduler;
@@ -18,7 +18,7 @@ describe('sleep', () => {
       const inputObservable = cold('a|', inputValues);
       const duration = 20;
 
-      const inputPorts = new LinkNodePorts(inputObservable, 'input');
+      const inputPorts = new LinkElementPorts(inputObservable, 'input');
       const sleepNode = Sleep.boot(duration);
       const outputPorts = sleepNode.getOutput(inputPorts);
       const output$ = outputPorts.getPort('output');
@@ -33,7 +33,7 @@ describe('sleep', () => {
       const inputObservable = cold('a - b - c|', inputValues);
       const duration = 20;
 
-      const inputPorts = new LinkNodePorts(inputObservable, 'input');
+      const inputPorts = new LinkElementPorts(inputObservable, 'input');
       const sleepNode = Sleep.boot(duration);
       const outputPorts = sleepNode.getOutput(inputPorts);
       const outputObservable = outputPorts.getPort('output');
@@ -44,7 +44,7 @@ describe('sleep', () => {
 
   it('should return EMPTY when requested port is not "output"', () => {
     const sleepNode = Sleep.boot(10);
-    const outputPorts = sleepNode.getOutput(new LinkNodePorts(of({}), 'input'));
+    const outputPorts = sleepNode.getOutput(new LinkElementPorts(of({}), 'input'));
     const outputObservable = outputPorts.getPort('notOutput');
 
     testScheduler.run(({ expectObservable }) => {
