+++ packages/core/src/V2/nodes/sleep.ts
@@ -1,12 +1,12 @@
 import { OperatorElementConfig } from '../circuitElement';
 import { delay } from 'rxjs/operators';
 import { Operator } from './operator';
-import { CreateOutputPort, NodePorts } from './nodePorts';
+import { CreateOutputPort, ElementPorts } from './elementPorts';
 
 export const Sleep: OperatorElementConfig = {
   boot: (param: unknown) => {
     const duration = Number(param);
-    let createSleepOutput: CreateOutputPort = (input) => new NodePorts(input.getPort('input').pipe(delay(duration)));
+    let createSleepOutput: CreateOutputPort = (input) => new ElementPorts(input.getPort('input').pipe(delay(duration)));
     return new Operator(createSleepOutput);
   }
 }
